requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
loguru==0.7.2
python-dotenv==1.0.1
sqlalchemy==2.0.39
psycopg2-binary==2.9.10